device: move event handler and statemachine to separate class/file
