+++ softhddevice.cpp
@@ -32,7 +32,6 @@ extern "C" {
 #include "audio.h"
 #include "codec_audio.h"
 #include "config.h"
-#include "event.h"
 #include "grab.h"
 #include "hardwaredevice.h"
 #include "jittertracker.h"
@@ -41,6 +40,7 @@ extern "C" {
 #include "pipreceiver.h"
 #include "softhddevice.h"
 #include "softhdosdprovider.h"
+#include "statemachine.h"
 #include "videorender.h"
 #include "videostream.h"
 
@@ -54,6 +54,7 @@ extern "C" {
 cSoftHdDevice::cSoftHdDevice(cSoftHdConfig *config)
 	: m_pConfig(config)
 {
+	m_pStateMachine = std::make_unique<cStateMachine>(this);
 }
 
 /**
@@ -87,7 +88,7 @@ bool cSoftHdDevice::Initialize(void)
 	// the following are deleted in the destructor
 	m_pSpuDecoder = new cDvbSpuDecoder();
 	m_pHardwareDevice = new cHardwareDevice();
-	m_pEventHandler = new cEventHandler(this);
+	m_pEventHandler = new cEventHandler(m_pStateMachine.get());
 
 	m_channelSwitchStartTime = std::chrono::steady_clock::now();
 	m_dataReceivedTime = m_channelSwitchStartTime;
@@ -105,7 +106,7 @@ bool cSoftHdDevice::Initialize(void)
 int cSoftHdDevice::Start(void)
 {
 	LOGDEBUG("device: %s", __FUNCTION__);
-	OnEventReceived(AttachEvent{});
+	TriggerEvent(AttachEvent{});
 
 	return true;
 }
@@ -117,7 +118,7 @@ void cSoftHdDevice::Stop(void)
 {
 	LOGDEBUG("device: %s", __FUNCTION__);
 	m_pPipHandler->Disable();
-	OnEventReceived(DetachEvent{});
+	TriggerEvent(DetachEvent{});
 }
 
 /*********************************************************************
@@ -325,26 +326,26 @@ bool cSoftHdDevice::SetPlayMode(ePlayMode play_mode)
 
 	// A new play mode arrived, attach first if we did detach because of an external player
 	if (m_externalPlayerActive) {
-		OnEventReceived(AttachEvent{});
+		TriggerEvent(AttachEvent{});
 		m_externalPlayerActive = false;
 	}
 
 	switch (play_mode) {
 	case pmNone:
-		OnEventReceived(StopEvent{});
+		TriggerEvent(StopEvent{});
 		break;
 	case pmAudioVideo:
 	case pmAudioOnly:
 	case pmAudioOnlyBlack:
 	case pmVideoOnly:
-		OnEventReceived(PlayEvent{});
+		TriggerEvent(PlayEvent{});
 		break;
 	case pmExtern_THIS_SHOULD_BE_AVOIDED:
 		// External players like mpv (vdr-plugin-mpv) want to acquire DRM/ALSA
 		// so we release it here and set a flag. As soon as the next SetPlayMode arrives
 		// we then can attach again before changing to the new playmode.
 		m_pPipHandler->Disable();
-		OnEventReceived(DetachEvent{});
+		TriggerEvent(DetachEvent{});
 		m_externalPlayerActive = true;
 		break;
 	default:
@@ -415,7 +416,7 @@ int cSoftHdDevice::PlayAudio(const uchar *data, int size, uchar id)
 	m_audioReassemblyBuffer.Push(pesPacket.GetPayload(), pesPacket.GetPayloadSize(), pesPacket.GetPts());
 
 	if (IsBufferingThresholdReached())
-		OnEventReceived(BufferingThresholdReachedEvent{});
+		TriggerEvent(BufferingThresholdReachedEvent{});
 
 	AVPacket *avpkt;
 	do {
@@ -612,7 +613,7 @@ void cSoftHdDevice::TrickSpeed(int speed, bool forward)
 		break;
 	}
 
-	OnEventReceived(TrickSpeedEvent{normalizedSpeed, speed != 0, forward});
+	TriggerEvent(TrickSpeedEvent{normalizedSpeed, speed != 0, forward});
 }
 
 /**
@@ -649,7 +650,7 @@ void cSoftHdDevice::Clear(void)
 	m_pAudio->ResetHwDelayBaseline();
 	FlushAudio();
 
-	SetState(BUFFERING);
+	m_pStateMachine->SetState(BUFFERING);
 
 	m_pRender->Resume();
 	m_pVideoStream->Resume();
@@ -664,7 +665,7 @@ void cSoftHdDevice::Play(void)
 {
 	cDevice::Play();
 
-	OnEventReceived(PlayEvent{});
+	TriggerEvent(PlayEvent{});
 }
 
 /**
@@ -675,7 +676,7 @@ void cSoftHdDevice::Freeze(void)
 	LOGDEBUG("device: %s:", __FUNCTION__);
 	cDevice::Freeze();
 
-	OnEventReceived(PauseEvent{});
+	TriggerEvent(PauseEvent{});
 }
 
 /**
@@ -693,7 +694,7 @@ void cSoftHdDevice::StillPicture(const uchar *data, int size)
 		return;
 	}
 
-	OnEventReceived(StillPictureEvent{data, size});
+	TriggerEvent(StillPictureEvent{data, size});
 }
 
 /**
@@ -926,7 +927,7 @@ void cSoftHdDevice::SetDisplayMode(int idx)
  */
 bool cSoftHdDevice::IsBufferingThresholdReached()
 {
-	if (m_state != BUFFERING)
+	if (m_pStateMachine->GetState() != BUFFERING)
 		return false;
 
 	bool audioHasInputPts = m_pAudio->HasInputPts();
@@ -1199,7 +1200,7 @@ void cSoftHdDevice::Detach(void)
 	}
 
 	m_pPipHandler->Disable();
-	OnEventReceived(DetachEvent{});
+	TriggerEvent(DetachEvent{});
 }
 
 /**
@@ -1217,7 +1218,7 @@ void cSoftHdDevice::Attach(void)
 		m_needsMakePrimary = false;
 	}
 
-	OnEventReceived(AttachEvent{});
+	TriggerEvent(AttachEvent{});
 }
 
 /**
@@ -1226,7 +1227,7 @@ void cSoftHdDevice::Attach(void)
 bool cSoftHdDevice::IsDetached(void) const
 {
 	std::lock_guard<std::mutex> lock(m_mutex);
-	return m_state == State::DETACHED;
+	return m_pStateMachine->GetState() == State::DETACHED;
 }
 
 /*********************************************************************
@@ -1538,286 +1539,18 @@ int cSoftHdDevice::GetBufferFillLevelThresholdMs() {
 }
 
 /*********************************************************************
- * cSoftHdDevice state machine and event handler  functions
+ * cSoftHdDevice state transitioning functions
  ********************************************************************/
 
 /**
- * Event handler for playback state transitions
+ * With this wrapper function, the device can directly act
+ * as an event reveiver.
  *
- * Processes events (Play, Pause, Stop, TrickSpeed, StillPicture) and performs
- * the appropriate state transitions based on the current state. The method halts
- * both display and decoding threads before processing the event and resumes them
- * afterwards to ensure thread-safe state transitions.
- *
- * @param event     The event to process (variant type containing specific event data)
+ * @param event           event to be executed
  */
-void cSoftHdDevice::OnEventReceived(const Event& event)
+void cSoftHdDevice::TriggerEvent(const Event &event)
 {
-	uint64_t startStateChange = cTimeMs::Now();
-	LOGDEBUG("STATE MACHINE: received %s", EventToString(event));
-	bool needsResume = false;
-
-#ifdef USE_GLES
-	// Lock the GL thread before the state machine lock, because cmdCopyBufferToOutputFb() calls
-	// cSoftHdDevice::OsdDrawARGB(), which itself locks the state machine mutex and we can end
-	// up in a deadlock then.
-	// We can safely unlock the thread again after the state change, because cSoftHdDevice::OsdDrawARGB()
-	// always tests if we are in detached mode and this new state is probably set then.
-	bool needsOglResume = false;
-	if (m_pOsdProvider && m_state != DETACHED)
-		needsOglResume = m_pOsdProvider->LockOpenGlThread();
-#endif
-
-	{ // locked state machine context
-	std::lock_guard<std::mutex> lock(m_mutex);
-
-	if (m_state != DETACHED) {
-		m_pRender->Halt();
-		m_pVideoStream->Halt();
-		needsResume = true;
-	}
-
-	auto invalid = [this, &event]() {
-		LOGWARNING("STATE MACHINE: Invalid event '%s' in state '%s' received", EventToString(event), StateToString(m_state));
-	};
-
-	switch (m_state) {
-		case State::DETACHED:
-			std::visit(overload{
-				[&invalid](const PlayEvent&) { invalid(); },
-				[&invalid](const PauseEvent&) { invalid(); },
-				[&invalid](const StopEvent&) { invalid(); },
-				[&invalid](const TrickSpeedEvent&) { invalid(); },
-				[&invalid](const StillPictureEvent&) { invalid(); },
-				[](const DetachEvent&) { /* ignore */ },
-				[this](const AttachEvent&) {
-					if (!m_forceDetached)
-						SetState(STOP);
-				},
-				[&invalid](const BufferUnderrunEvent&) { invalid(); },
-				[&invalid](const BufferingThresholdReachedEvent&) { invalid(); },
-				[&invalid](const ScheduleResyncAtPtsMsEvent&) { invalid(); },
-				[&invalid](const ResyncEvent&) { invalid(); },
-				[&invalid](const DisplayChangeEvent&) { invalid(); },
-			}, event);
-			needsResume = false;
-			break;
-		case State::STOP:
-			std::visit(overload{
-				[this](const PlayEvent&) {
-					m_pAudio->LazyInit();
-					m_pAudio->SetVolume((m_volume * 1000) / 255);
-					SetState(BUFFERING);
-					m_pRender->ResetFrameCounter();
-					m_pVideoStream->ResetFilterThreadNeededCheck();
-				},
-				[&invalid](const PauseEvent&) { invalid(); },
-				[&invalid](const StopEvent&) { invalid(); },
-				[&invalid](const TrickSpeedEvent&) { invalid(); },
-				[this](const StillPictureEvent& s) {
-					HandleStillPicture(s.data, s.size);
-				},
-				[this, &needsResume](const DetachEvent&) {
-					SetState(DETACHED);
-					needsResume = false;
-				},
-				[&invalid](const AttachEvent&) { invalid(); },
-				[&invalid](const BufferUnderrunEvent&) { invalid(); },
-				[&invalid](const BufferingThresholdReachedEvent&) { invalid(); },
-				[&invalid](const ScheduleResyncAtPtsMsEvent&) { invalid(); },
-				[&invalid](const ResyncEvent&) { invalid(); },
-				[this](const DisplayChangeEvent& d) {
-					HandleDisplayModeChange(d.mode);
-				},
-			}, event);
-			break;
-		case State::BUFFERING:
-			std::visit(overload{
-				[this](const PlayEvent&) {
-					// ignore
-				},
-				[this](const PauseEvent&) {
-					// ignore
-				},
-				[this](const StopEvent&) {
-					SetState(STOP);
-				},
-				[this](const TrickSpeedEvent& t) {
-					// abort buffering and proceed with trick speed immediately, because trick speed shall be as fast and as demanded as possible
-					SetState(PLAY);
-					m_pRender->SetTrickSpeed(t.speed, t.active, t.forward);
-					SetState(TRICK_SPEED);
-				},
-				[this](const StillPictureEvent& s) {
-					HandleStillPicture(s.data, s.size);
-				},
-				[this, &needsResume](const DetachEvent&) {
-					SetState(DETACHED);
-					needsResume = false;
-				},
-				[&invalid](const AttachEvent&) { invalid(); },
-				[&invalid](const BufferUnderrunEvent&) { invalid(); },
-				[this](const BufferingThresholdReachedEvent&) {
-					bool receivedAudio = m_pAudio->HasInputPts();
-					bool receivedVideo = m_pVideoStream->HasInputPts();
-
-					if (receivedAudio && receivedVideo) {
-						m_playbackMode = AUDIO_AND_VIDEO;
-						int64_t firstAudioPtsMs = GetFirstAudioPtsMsToPlay();
-						int64_t firstVideoPtsMs = GetFirstVideoPtsMsToPlay();
-						// store the first PTSes beforehand, because dropping samples/frames will change the output of GetFirst*PtsMsToPlay()
-						m_pAudio->DropSamplesOlderThanPtsMs(firstAudioPtsMs);
-						m_pRender->SchedulePlaybackStartAtPtsMs(firstVideoPtsMs);
-					} else if (receivedAudio) {
-						LOGDEBUG("device: audio only detected");
-						m_playbackMode = AUDIO_ONLY;
-						m_pAudio->DropSamplesOlderThanPtsMs(m_pAudio->GetOutputPtsMs());
-					} else if (receivedVideo) {
-						LOGDEBUG("device: video only detected");
-						m_playbackMode = VIDEO_ONLY;
-						m_pRender->SchedulePlaybackStartAtPtsMs(m_pRender->GetOutputPtsMs());
-					} else {
-						// Sometimes a DeviceClear() can jump in between signalling the ThresholdReachedEvent
-						// and progressing it, e.g. the video thread signals the ThresholdReached, VDR sends a DeviceClear()
-						// and OnEventReceived wants to process the ThresholdReachedEvent with empty buffers.
-						LOGDEBUG("device: buffering threshold reached and no a/v available, keep BUFFERING state");
-					}
-
-					if (receivedAudio || receivedVideo)
-						SetState(PLAY);
-				},
-				[this](const ScheduleResyncAtPtsMsEvent& s) {
-					SetState(PLAY);
-					m_pRender->ScheduleResyncAtPtsMs(s.pts);
-				},
-				[&invalid](const ResyncEvent&) { invalid(); },
-				[this](const DisplayChangeEvent& d) {
-					HandleDisplayModeChange(d.mode);
-				},
-			}, event);
-			break;
-		case State::PLAY:
-			std::visit(overload{
-				[this](const PlayEvent&) {
-					// resume from pause
-					int audioBehindVideoByMs;
-					switch (m_playbackMode) {
-						case AUDIO_ONLY:
-							m_pAudio->SetHwDelayBaseline();
-							m_pAudio->SetPaused(false);
-							break;
-						case AUDIO_AND_VIDEO:
-							audioBehindVideoByMs = m_pRender->GetOutputPtsMs() - m_pAudio->GetOutputPtsMs() - m_pConfig->ConfigVideoAudioDelayMs;
-							m_pAudio->SetHwDelayBaseline();
-							if (audioBehindVideoByMs > 0) {
-								m_pAudio->DropSamplesOlderThanPtsMs(m_pAudio->GetOutputPtsMs() + audioBehindVideoByMs);
-								m_pAudio->SetPaused(false);
-							} else
-								m_pRender->SetScheduleAudioResume(true);
-
-							// fallthrough
-						case VIDEO_ONLY:
-							m_pRender->SetStillpicture(false);
-							m_pRender->SetPlaybackPaused(false);
-							break;
-						case NONE:
-							LOGFATAL("STATE MACHINE: play event in PLAY state with NONE playback mode. This is a bug.");
-							break;
-					}
-				},
-				[this](const PauseEvent&) {
-					if (m_playbackMode == AUDIO_AND_VIDEO)
-						m_pRender->ScheduleVideoPlaybackPauseAt(m_pAudio->GetOutputPtsMs() - m_pConfig->ConfigVideoAudioDelayMs);
-					else
-						m_pRender->SetPlaybackPaused(true);
-
-					m_pAudio->SetPaused(true);
-					m_pAudio->ResetHwDelayBaseline();
-				},
-				[this](const StopEvent&) {
-					SetState(STOP);
-				},
-				[this](const TrickSpeedEvent& t) {
-					m_pRender->SetTrickSpeed(t.speed, t.active, t.forward);
-					SetState(TRICK_SPEED);
-				},
-				[this](const StillPictureEvent& s) {
-					HandleStillPicture(s.data, s.size);
-				},
-				[this, &needsResume](const DetachEvent&) {
-					SetState(DETACHED);
-					needsResume = false;
-				},
-				[&invalid](const AttachEvent&) { invalid(); },
-				[this](const BufferUnderrunEvent&) {
-					SetState(BUFFERING);
-				},
-				[&invalid](const BufferingThresholdReachedEvent&) {
-					// ignore
-				},
-				[this](const ScheduleResyncAtPtsMsEvent& s) {
-					m_pRender->ScheduleResyncAtPtsMs(s.pts);
-				},
-				[this](const ResyncEvent&) {
-					SetState(BUFFERING);
-				},
-				[this](const DisplayChangeEvent& d) {
-					HandleDisplayModeChange(d.mode);
-				},
-			}, event);
-			break;
-		case State::TRICK_SPEED:
-			std::visit(overload{
-				[this](const PlayEvent&) {
-					SetState(PLAY);
-				},
-				[this](const PauseEvent&) {
-					m_pRender->SetPlaybackPaused(true);
-					m_pAudio->SetPaused(true);
-					m_pAudio->ResetHwDelayBaseline();
-				},
-				[this](const StopEvent&) {
-					SetState(STOP);
-				},
-				[this](const TrickSpeedEvent& t) {
-					// resume from pause, or change trick speed direction/speed
-					m_pRender->SetTrickSpeed(t.speed, t.active, t.forward);
-					m_pRender->SetPlaybackPaused(false);
-				},
-				[this](const StillPictureEvent& s) {
-					HandleStillPicture(s.data, s.size);
-				},
-				[this, &needsResume](const DetachEvent&) {
-					SetState(DETACHED);
-					needsResume = false;
-				},
-				[&invalid](const AttachEvent&) { invalid(); },
-				[this](const BufferUnderrunEvent&) {
-					// ignore during trick speed. Fast forward/reverse as fast and as demanded as possible
-				},
-				[&invalid](const BufferingThresholdReachedEvent&) { invalid(); },
-				[&invalid](const ScheduleResyncAtPtsMsEvent&) { invalid(); },
-				[&invalid](const ResyncEvent&) { invalid(); },
-				[this](const DisplayChangeEvent& d) {
-					HandleDisplayModeChange(d.mode);
-				},
-			}, event);
-			break;
-	}
-
-	if (needsResume) {
-		m_pVideoStream->Resume();
-		m_pRender->Resume();
-	}
-
-	} // end of locked state machine context
-#ifdef USE_GLES
-	if (m_pOsdProvider && needsOglResume)
-		m_pOsdProvider->UnlockOpenGlThread();
-#endif
-
-	uint64_t stopStateChange = cTimeMs::Now();
-	LOGDEBUG("STATE MACHINE: state change done in %d ms", (int)(stopStateChange - startStateChange));
+	m_pStateMachine->OnEventReceived(event);
 }
 
 /**
@@ -1828,7 +1561,8 @@ void cSoftHdDevice::OnEventReceived(const Event& event)
  *
  * @param state         state being left
  */
-void cSoftHdDevice::OnLeavingState(State state) {
+void cSoftHdDevice::LeaveState(State state)
+{
 	switch (state) {
 		case PLAY:
 			m_pRender->SchedulePlaybackStartAtPtsMs(AV_NOPTS_VALUE);
@@ -1882,7 +1616,8 @@ void cSoftHdDevice::OnLeavingState(State state) {
  *
  * @param state         state being entered
  */
-void cSoftHdDevice::OnEnteringState(State state) {
+void cSoftHdDevice::EnterState(State state)
+{
 	switch (state) {
 		case BUFFERING:
 			m_pAudio->ResetHwDelayBaseline();
@@ -1960,70 +1695,190 @@ void cSoftHdDevice::OnEnteringState(State state) {
 }
 
 /**
- * Sets the device into the given state.
+ * Pause the rendering and decoder thread.
+ */
+void cSoftHdDevice::HaltVideoThreads(void)
+{
+	m_pRender->Halt();
+	m_pVideoStream->Halt();
+}
+
+/**
+ * Resume the rendering and decoder thread.
+ */
+void cSoftHdDevice::ResumeVideoThreads(void)
+{
+	m_pVideoStream->Resume();
+	m_pRender->Resume();
+}
+
+#ifdef USE_GLES
+/**
+ * Pause the OpenGL worker thread.
+ */
+bool cSoftHdDevice::HaltOpenGlThread(void)
+{
+	return m_pOsdProvider && m_pOsdProvider->LockOpenGlThread();
+}
+
+/**
+ * Resume the OpenGL worker thread.
+ */
+void cSoftHdDevice::ResumeOpenGlThread(void)
+{
+	if (m_pOsdProvider)
+		m_pOsdProvider->UnlockOpenGlThread();
+}
+#endif
+
+/**
+ * Returns true, the a detached plugin start was forced.
+ */
+bool cSoftHdDevice::IsDetachForced(void)
+{
+	return m_forceDetached;
+}
+
+/**
+ * Init the audio lazily.
  *
- * @param newState       new state
+ * @param setVolume           if true, reset the current volume
  */
-void cSoftHdDevice::SetState(State newState)
+void cSoftHdDevice::InitAudio(bool setVolume)
 {
-	if (m_state != newState) {
-		LOGDEBUG("STATE MACHING: Preparing to leave state %s", StateToString(m_state));
-		OnLeavingState(m_state);
-		LOGDEBUG("STATE MACHING: Changing state %s -> %s", StateToString(m_state), StateToString(newState));
-		m_state = newState;
-		OnEnteringState(m_state);
-		LOGDEBUG("STATE MACHING: State changed to %s", StateToString(m_state));
+	m_pAudio->LazyInit();
+	if (setVolume)
+		m_pAudio->SetVolume((m_volume * 1000) / 255);
+}
+
+/**
+ * Reset the video deinterlace filter
+ */
+void cSoftHdDevice::ResetVideoFilter(void)
+{
+	m_pRender->ResetFrameCounter();
+	m_pVideoStream->ResetFilterThreadNeededCheck();
+}
+
+/**
+ * Set the trickspeed in the renderer
+ *
+ * @param speed           trickspeed factor
+ * @param active          true, if trickspeed is active
+ * @param forward         true, if this is forward trickspeed
+ */
+void cSoftHdDevice::SetTrickSpeed(double speed, bool active, bool forward)
+{
+	m_pRender->SetTrickSpeed(speed, active, forward);
+}
+
+/**
+ * Schedules the playback start
+ *
+ * Drops audio data if necessary to start in sync and schedules
+ * the playback start at the common pts.
+ *
+ * @retval true                  playback started
+ * @retval false                 playback was not started
+ */
+bool cSoftHdDevice::SchedulePlaybackStart(void)
+{
+	bool receivedAudio = m_pAudio->HasInputPts();
+	bool receivedVideo = m_pVideoStream->HasInputPts();
+
+	if (receivedAudio && receivedVideo) {
+		m_playbackMode = AUDIO_AND_VIDEO;
+		int64_t firstAudioPtsMs = GetFirstAudioPtsMsToPlay();
+		int64_t firstVideoPtsMs = GetFirstVideoPtsMsToPlay();
+
+		// store the first PTSes beforehand, because dropping samples/frames will change the output of GetFirst*PtsMsToPlay()
+		m_pAudio->DropSamplesOlderThanPtsMs(firstAudioPtsMs);
+		m_pRender->SchedulePlaybackStartAtPtsMs(firstVideoPtsMs);
+	} else if (receivedAudio) {
+		LOGDEBUG("device: audio only detected");
+		m_playbackMode = AUDIO_ONLY;
+		m_pAudio->DropSamplesOlderThanPtsMs(m_pAudio->GetOutputPtsMs());
+	} else if (receivedVideo) {
+		LOGDEBUG("device: video only detected");
+		m_playbackMode = VIDEO_ONLY;
+		m_pRender->SchedulePlaybackStartAtPtsMs(m_pRender->GetOutputPtsMs());
+	} else {
+		// Sometimes a DeviceClear() can jump in between signalling the ThresholdReachedEvent
+		// and progressing it, e.g. the video thread signals the ThresholdReached, VDR sends a DeviceClear()
+		// and OnEventReceived wants to process the ThresholdReachedEvent with empty buffers.
+		LOGDEBUG("device: buffering threshold reached and no a/v available, keep BUFFERING state");
+		return false;
+	}
+
+	return true;
+}
+
+/**
+ * Schedule aa a/v resync in the render thread at the given pts
+ *
+ * @param pts            resync, when the rendered video frame reaches this pts
+ */
+void cSoftHdDevice::ScheduleResyncAtPtsMs(int64_t pts)
+{
+	m_pRender->ScheduleResyncAtPtsMs(pts);
+}
+
+/**
+ * Resume playback from pause state
+ *
+ * Other than ResumePlayback() this also syncs a/v (by dropping audio)
+ */
+void cSoftHdDevice::ResumeFromPause(void)
+{
+	// resume from pause
+	int audioBehindVideoByMs;
+
+	switch (m_playbackMode) {
+		case AUDIO_ONLY:
+			m_pAudio->SetHwDelayBaseline();
+			m_pAudio->SetPaused(false);
+			break;
+		case AUDIO_AND_VIDEO:
+			audioBehindVideoByMs = m_pRender->GetOutputPtsMs() - m_pAudio->GetOutputPtsMs() - m_pConfig->ConfigVideoAudioDelayMs;
+			m_pAudio->SetHwDelayBaseline();
+			if (audioBehindVideoByMs > 0) {
+				m_pAudio->DropSamplesOlderThanPtsMs(m_pAudio->GetOutputPtsMs() + audioBehindVideoByMs);
+				m_pAudio->SetPaused(false);
+			} else
+				m_pRender->SetScheduleAudioResume(true);
+
+			// fallthrough
+		case VIDEO_ONLY:
+			m_pRender->SetStillpicture(false);
+			m_pRender->SetPlaybackPaused(false);
+			break;
+		case NONE:
+			LOGFATAL("STATE MACHINE: play event in PLAY state with NONE playback mode. This is a bug.");
+			break;
 	}
 }
 
 /**
- * Create and start the event handler thread
+ * Pause the playback
+ *
+ * @param schedulePause               if true, pause at the current audio pts
+ *                                    otherwise pause immediately
  */
-cEventHandler::cEventHandler(cSoftHdDevice *device)
-	: cThread("event handler"),
-	  m_pDevice(device),
-	  m_pEventReceiver(device)
+void cSoftHdDevice::PausePlayback(bool schedulePause)
 {
-	Start();
+	if (schedulePause && m_playbackMode == AUDIO_AND_VIDEO)
+		m_pRender->ScheduleVideoPlaybackPauseAt(m_pAudio->GetOutputPtsMs() - m_pConfig->ConfigVideoAudioDelayMs);
+	else
+		m_pRender->SetPlaybackPaused(true);
+
+	m_pAudio->SetPaused(true);
+	m_pAudio->ResetHwDelayBaseline();
 }
 
 /**
- * Stop and delete the event handler thread
+ * Resume playback again
  */
-cEventHandler::~cEventHandler(void)
+void cSoftHdDevice::ResumePlayback(void)
 {
-	Cancel(2);
-}
-
-/**
- * Periodically send events in the queue to the final event receiver/handler
- */
-void cEventHandler::Action(void)
-{
-	LOGDEBUG("device: event queue handler thread started");
-
-	while (Running()) {
-		std::vector<Event> local;
-		{
-			std::lock_guard<std::mutex> lock(m_mutex);
-			local.swap(m_eventQueue);
-		}
-
-		for (auto &event : local)
-			m_pEventReceiver->OnEventReceived(event);
-
-		usleep(10000);
-
-	}
-
-	LOGDEBUG("device: event queue handler thread stopped");
-}
-
-/**
- * Add an event to the queue
- */
-void cEventHandler::AddEvent(Event event)
-{
-	std::lock_guard<std::mutex> lock(m_mutex);
-	m_eventQueue.push_back(event);
+	m_pRender->SetPlaybackPaused(false);
 }
